Import now works again, ids in the schema are copied, only ids in the lexicon are created anew.
Deleting a lexicon no longer crashes or removes the name of a lexicon in the list.

Klingon import slightly improved.

Fix bug in creation of standard view in the import process.

Import now returns a <lexicon> element containing the info the UI needs.

Desktop pom reverted back to basex 6.7 (should be 6.7.1 because that contains a fix we need, but desktop isn't working yet anyway, so I guess, to keep Jenkins happy, we'll just use a 6.7 dependency for now).